+++ src/main/java/demo/springapi/customerapi/repository/CustomerRepository.java
@@ -4,4 +4,4 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import demo.springapi.customerapi.entity.Customer;
 
-public interface CustomerRepository extends JpaRepository<Customer, Integer> {}
+public interface CustomerRepository extends JpaRepository<Customer, Long> {}
